+++ Examples/Scripts/Python/Setup/update_build_configuration.py
@@ -1,3 +1,7 @@
+# /// script
+# dependencies = ["lxml"]
+# ///
+
 import os
 import multiprocessing
 from lxml import etree